# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Secure Form API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/form": {
            "post": {
                "summary": "Submit form",
                "operationId": "ac399e5aa7056b46a2a118392c8cf376",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "file_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Form submitted successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/files/upload": {
            "post": {
                "tags": [
                    "File Upload"
                ],
                "summary": "Upload a file",
                "operationId": "6fa43dfe477c31a48e259170f7556825",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "file": {
                                        "description": "File to upload",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "File uploaded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "File uploaded successfully"
                                        },
                                        "file": {
                                            "description": "Uploaded file details",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "An error occurred during file upload",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "An error occurred during file upload. Please try again later."
                                        },
                                        "details": {
                                            "type": "string",
                                            "example": "Error details here."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "File Upload",
            "description": "File Upload"
        }
    ]
}